backend/requirements.txt
bcrypt
fastapi
motor
openai
pydantic
pyjwt
pymongo
uvicorn